full rm hashmap
